requirements-be.txt
fastapi
uvicorn
pydantic
textblob
joblib
scikit-learn
pandas
numpy
chromadb
google-generativeai
python-dotenv
